+++ plugin_base/src/plugin_base/plugin_base/topo/section.cpp
@@ -14,6 +14,21 @@ param_section::validate(plugin_topo const& plugin, module_topo const& module, in
   assert(this->index == index_);
   auto always_visible = [&module](int p) { return !module.params[p].gui.bindings.visible.is_bound(); };
   auto include = [this, &module](int p) { return module.params[p].gui.visible && module.params[p].gui.section == this->index; };
+
+  bool is_one_grid_param = false;
+  for(int p = 0; p < module.params.size(); p++)
+    if (include(p))
+    {
+      if (module.params[p].gui.layout == param_layout::single_grid)
+      {
+        assert(!is_one_grid_param);
+        is_one_grid_param = true;
+      }
+    }
+
+  // validation is handled elsewhere
+  if(is_one_grid_param) return;
+
   gui.dimension.validate(gui.cell_split, 
     vector_map(module.params, [](auto const& p){ return p.gui.position; }), 
     vector_map(module.params, [](auto const& p){ return p.gui.label.contents; }),